seperated the views and created the api_documentation
