+++ backend/genfit_django/api/urls.py
@@ -3,7 +3,7 @@ from rest_framework.routers import DefaultRouter
 from . import views
 from .separate_views import fitness_goals
 from .separate_views import profile
-from .views import ForumViewSet, ThreadViewSet
+from .separate_views.forum import ForumViewSet, ThreadViewSet
 
 
 urlpatterns = [
@@ -32,3 +32,5 @@ urlpatterns = [
 router = DefaultRouter()
 router.register(r'forums', ForumViewSet)
 router.register(r'threads', ThreadViewSet)
+
+urlpatterns += router.urls
